Move stamina bar to RscTitles
