+++ game/configs/mission/ui/inc/idcs.inc
@@ -114,3 +114,7 @@
 #define VGM_IDC_PROGRESSBAR_PROGRESSBAR 1000
 #define VGM_IDC_PROGRESSBAR_TITLE 1001
 #define VGM_IDC_PROGRESSBAR_DRAWHANDLER 1002
+
+// Stamina Bar
+#define VGM_IDC_STAMINABAR_CONTAINER 1000
+#define VGM_IDC_STAMINABAR_BAR 1001
