+++ game/functions/functions_client.hpp
@@ -96,6 +96,7 @@ class vgm_c
         class displayMissions {};
         class displayMissionsTargets {};
         class displayAbilityCooldown {};
+        class displayStaminaBar {};
     };
 
     class missions
