+++ game/functions/systems/stamina/client/fn_stamina_postInit.sqf
@@ -2,7 +2,7 @@
     File: fn_stamina_postInit.sqf
     Author: Savage Game Design
     Date: 2023-08-18
-    Last Update: 2023-08-18
+    Last Update: 2023-08-22
     Public: No
 
     Description:
@@ -12,3 +12,5 @@
 if (!hasInterface) exitWith {};
 
 player call vgm_c_fnc_stamina_unitInit;
+
+"vgm_stamina_bar" cutRsc ["vgm_RscStaminaBar", "PLAIN", -1, false];
